+++ src/customMenu.ts
@@ -1,0 +1,14 @@
+import { createCustomMenu } from "./sheet/createCustomMenu";
+
+/**
+ * Custom menu
+ */
+declare const global: {
+  [x: string]: () => void;
+};
+
+function onOpen(): void {
+  createCustomMenu();
+}
+
+global.onOpen = onOpen;
